# Improve rendering performance of query result panel