--- a/package.json
+++ b/package.json
@@ -40,6 +40,7 @@
     "react-dom": "19.0.0",
     "react-markdown": "^10.1.0",
     "react-resizable-panels": "^2.1.7",
+    "throttleit": "^2.1.0",
     "valibot": "^1.0.0"
   },
   "devDependencies": {