@ptgolden 2025-07-01
Fixes #34